tunnel: fix path_rewrites feature not handled in transport layer
This commit re-implements the path_rewrites feature from the Envoy backend version of TNG in the native implementation.

To support this feature, which requires HTTP request path extraction, we introduced an `HttpRequestInspector`. This module can sniff out HTTP requests from any `AsyncRead` stream without altering the data. It's designed to be efficient—unlike a full HTTP proxy, it doesn't reconstruct the HTTP request, but rather inspects just enough of the stream to obtain the necessary HTTP headers before continuing to forward raw TCP data bidirectionally.

The extracted path is then processed by the `PathRewriteGroup` module for path rewriting, and used as one of the inputs for the HTTP transport layer. Additionally, we now extract the correct authority (`host:port`), allowing the HTTP transport layer to send requests with the appropriate `Host` header. Previously, the Host header was often incorrectly set to `{ip}:{port}`.

Signed-off-by: Kun Lai <laikun@linux.alibaba.com>