tests/netns: change ping 8.8.8.8 to curl for check internet access
The ping (ICMP echo) traffic to the internet network seems got dropped
in github action which causing the CI failure. So let's change to use
curl to check network access.

Another problem is that in the container the dns server is set to a
localhost address:

```
nameserver 127.0.0.11
```
So in this commit we also set 8.8.8.8 as dns server in /etc/resolv.conf.

Signed-off-by: Kun Lai <laikun@linux.alibaba.com>