tng-wasm: remove streaming request support for better compatibility
Due to limited browser support and the requirement for HTTPS environment
for streaming requests (duplex: 'half'), this commit removes the streaming
request functionality for the wasm target, where the HTTP request ciphertext
is now converted to bytes before being sent to the upstream in one go.

This change improves compatibility across different browsers and removes
the HTTPS requirement that was limiting adoption in many environments.

Ref: https://developer.chrome.com/docs/capabilities/web-apis/fetch-streaming-requests

Signed-off-by: Kun Lai <laikun@linux.alibaba.com>